todo app done

+++ app/src/main/java/com/example/jetpackcompose_material2_demo/ui/update_note/UpdateNoteScreen.kt
@@ -2,7 +2,10 @@ package com.example.jetpackcompose_material2_demo.ui.update_note
 
 import android.util.Log
 import android.widget.Toast
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -10,8 +13,10 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.AlertDialog
 import androidx.compose.material.OutlinedTextField
 import androidx.compose.material.Text
+import androidx.compose.material.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -20,7 +25,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.dp
@@ -47,6 +54,7 @@ fun UpdateNoteScreen(navController: NavHostController, id: String) {
     val viewModel: UpdateNoteViewModel = hiltViewModel()
     val coroutineScope: CoroutineScope = rememberCoroutineScope()
     val context = LocalContext.current
+    val dialogueState = remember { mutableStateOf(false) }
 
     val noteModelState = viewModel.noteModel.collectAsState(initial = UpdateNoteState.Loading)
     when (val result = noteModelState.value) {
@@ -59,27 +67,59 @@ fun UpdateNoteScreen(navController: NavHostController, id: String) {
 
             var getNoteModel by rememberSaveable { mutableStateOf(result.model) }
 
-            Column(modifier = Modifier
-                .fillMaxSize()
-                .verticalScroll(state = rememberScrollState())) {
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .verticalScroll(state = rememberScrollState())
+            ) {
+
+                if (dialogueState.value) {
+                    AlertDialog(
+                        modifier = Modifier.fillMaxWidth(),
+                        onDismissRequest = { /*TODO*/ },
+                        buttons = {
+                            Box(
+//                                horizontalArrangement = Arrangement.SpaceEvenly,
+                                modifier = Modifier.fillMaxWidth(),
+                            ) {
+                                TextButton(
+                                    onClick = { dialogueState.value = false },
+                                    modifier = Modifier.align(Alignment.CenterStart).padding(start = 16.dp)
+                                ) {
+                                    Text(text = "Cancel", color = Color.Black)
+                                }
+                                TextButton(onClick = {
+
+                                    dialogueState.value = false
+
+                                    val getId = viewModel.deleteNote(id.toInt())
+                                    coroutineScope.launch(Dispatchers.Main) {
+                                        getId.collect { id ->
+                                            if (id > 0) {
+                                                navController.popBackStack()
+                                                Toast.makeText(
+                                                    context, "Note removed successfully",
+                                                    Toast.LENGTH_SHORT
+                                                ).show()
+                                            }
+                                        }
+                                    }
+                                }, modifier = Modifier.align(Alignment.CenterEnd)) {
+                                    Text(text = "Yes", color = Color.Red)
+                                }
+                            }
+                        },
+                        title = { Text(text = "Remove Note") },
+                        text = { Text(text = "Are you sure you want to remove note ?") })
+                }
 
                 AppBarC(
                     title = "Update Note",
                     showBackArrow = true,
                     showDeleteNoteIcon = true,
                     onDeleteClick = {
-                        val getId = viewModel.deleteNote(id.toInt())
-                        coroutineScope.launch(Dispatchers.Main) {
-                            getId.collect { id ->
-                                if (id > 0) {
-                                    navController.popBackStack()
-                                    Toast.makeText(
-                                        context, "Note removed successfully",
-                                        Toast.LENGTH_SHORT
-                                    ).show()
-                                }
-                            }
-                        }
+
+                        dialogueState.value = true
                     },
                     showSubmitIcon = getNoteModel.description.isNotBlank(),
                     onSubmitClick = {
@@ -204,7 +244,8 @@ fun UpdateNoteScreen(navController: NavHostController, id: String) {
 
         is UpdateNoteState.Error -> {
             Log.e("UpdateNoteScreen", "Error: ${result.exception}")
-            Toast.makeText(context, "Something went wrong in update note", Toast.LENGTH_SHORT).show()
+            Toast.makeText(context, "Something went wrong in update note", Toast.LENGTH_SHORT)
+                .show()
         }
     }
 }
